Move State Class into Base; Rename SetandState to Responder
