+++ insteon_mqtt/device/Base.py
@@ -9,6 +9,7 @@ import os.path
 from .MsgHistory import MsgHistory
 from ..Address import Address
 from ..CommandSeq import CommandSeq
+from ..Signal import Signal
 from .. import db
 from .. import handler
 from .. import log
@@ -119,6 +120,14 @@ class Base:
         # Special callback to run when receiving a broadcast clean up.  See
         # scene() for details.
         self.broadcast_reason = ""
+
+        # Used for internally tracking the device state
+        self._is_on = False
+        self._level = 0x00
+
+        # Support dimmer style signals and motion on/off style signals.
+        # API:  func(Device, int level, on_off.Mode mode, str reason)
+        self.signal_state = Signal()
 
         # Map (mqtt) commands mapped to methods calls.  These are handled in
         # run_command().  Derived classes can add more commands to the dict
@@ -1009,10 +1018,45 @@ class Base:
     #-----------------------------------------------------------------------
     def _set_state(self, is_on=None, level=None, group=None,
                    mode=on_off.Mode.NORMAL, reason=""):
-        # Implemented by functions.State.  Raises the question of if that
-        # function should be moved here, and disabled for those few devices
-        # that don't support it
-        pass
+        """Update the device level or on/off state.
+
+        This will change the internal state and emit the state changed
+        signals.  It is called by whenever we're informed that the device has
+        changed state.
+
+        Args:
+          is_on (bool):  True if the switch is on, False if it isn't.
+          level (int): The new device level in the range [0,255].  0 is off.
+          group (int): The group to which this applies
+          mode (on_off.Mode): The type of on/off that was triggered (normal,
+               fast, etc).
+          reason (str):  This is optional and is used to identify why the
+                 command was sent. It is passed through to the output signal
+                 when the state changes - nothing else is done with it.
+        """
+        LOG.info("Setting device %s on %s level %s %s %s", self.label, is_on,
+                 level, mode, reason)
+        self._cache_state(group, is_on, level, reason)
+
+        self.signal_state.emit(self, is_on=is_on, level=level, mode=mode,
+                               button=group, reason=reason)
+
+    #-----------------------------------------------------------------------
+    def _cache_state(self, group, is_on, level, reason):
+        """Cache the State of the Device
+
+        Used to help with the unique device functions.
+
+        Args:
+          group (int): The group which this applies
+          is_on (bool): Whether the device is on.
+          level (int): The new device level in the range [0,255].  0 is off.
+          reason (str): Reason string to pass around.
+        """
+        if is_on is not None:
+            self._is_on = is_on
+        if level is not None:
+            self._level = level
 
     #-----------------------------------------------------------------------
     def handle_received(self, msg):
