+++ insteon_mqtt/device/functions/__init__.py
@@ -19,7 +19,6 @@ the list of available functions to inherit from.
 
 from ..Base import Base
 from .Scene import Scene
-from .SetAndState import SetAndState
-from .State import State
+from .Responder import Responder
 from .Backlight import Backlight
 from .DimmerFuncs import DimmerFuncs
